# [Attention] Optimize FlashInfer MetadataBuilder Build call