## Essential Elements of an Effective PR Description Checklist
- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.

## Purpose

Flash infer prefers host side CPU buffers in many cases, example: https://github.com/flashinfer-ai/flashinfer/blob/3c40456effae8b9c5b1a11c0d1e0594295b1a312/flashinfer/prefill.py#L1430-L1436

So we pass host side buffers (since https://github.com/vllm-project/vllm/pull/20466 we now have access to these) to reduce D2H transfers.

Trace from main showing D2H transfers in `plan`

<img width="1565" height="310" alt="image" src="https://github.com/user-attachments/assets/48a4fb10-3579-4e2f-b791-7c264ad1e944" />

## Test Plan

## Test Result

### Accuracy Results

```
VLLM_ATTENTION_BACKEND=FLASHINFER lm_eval --model vllm --model_args pretrained=met
a-llama/Meta-Llama-3-8B-Instruct --tasks gsm8k --batch_size auto
...
INFO 07-17 20:33:43 [cuda.py:253] Using FlashInfer backend on V1 engine.
...
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7536|±  |0.0119|
|     |       |strict-match    |     5|exact_match|↑  |0.7551|±  |0.0118|
```

### Benchmark Results

**Benchmark Command:**
```bash
python benchmarks/benchmark_throughput.py --model meta-llama/Llama-3.2-3B-Instruct --dataset-name random --input-len 256 --output-len 128 --num-prompts <N> --seed 42
```

**Results** (3 runs per condition, mean ± standard error):

| num-prompts | Main Branch (req/s) | This PR (req/s) |
|-------------|---------------------|------------------------|
| 1           | 1.58 ± 0.06         | 1.90 ± 0.03           |
| 8           | 13.06 ± 0.11        | 14.32 ± 0.21          |
| 16          | 26.00 ± 0.07        | 28.74 ± 0.13          |
| 32          | 47.84 ± 0.57        | 46.53 ± 1.57          |
| 64          | 76.14 ± 0.45        | 81.43 ± 3.43          |
| 128         | 116.99 ± 6.10       | 127.78 ± 7.50         |
| 256         | 164.45 ± 6.12       | 177.70 ± 3.88         |

*Tested on NVIDIA B200 GPU with meta-llama/Llama-3.2-3B-Instruct (256→128 tokens)* 

## (Optional) Documentation Update

<!--- pyml disable-next-line no-emphasis-as-heading -->