@LucasWilkinson 2025-07-18
> Could we still pass the device tensors to Flashinfer's plan() rather than host tensors? Because we might want to support full cudagraph of Flashinfer in the future (currently implemented in #20059 in rough), which requires managing device-side persistent buffers that can be reused across different decode wrappers. Here, one decode wrapper corresponds to a runtime shape that needs to be captured.

If you look in FlashInfer's `BatchDecodeWithPagedKVCacheWrapper` you'll see the buffers get copied in the cudagraph pass regardless: https://github.com/flashinfer-ai/flashinfer/blob/1e9a41ad7f0efc5989bb0a2bf7e954902c8c73af/flashinfer/decode.py#L892-L910 and will get copied to the host: https://github.com/flashinfer-ai/flashinfer/blob/1e9a41ad7f0efc5989bb0a2bf7e954902c8c73af/flashinfer/decode.py#L925-L926

> Also, if we pass the host tensors to the wrapper, it seems that H2D transfers still exist. 

Yes; however H2D transfers are preferred over D2H as they can be done in a non-blocking fashion and do force synchronization with GPU. For the build call we are trying to optimize the CPU overhead so the fire-and-forget nature of the H2D transfers is better then depending on D2H transfer.

> If I remember correctly, Sglang's implementation overrides the plan functions that still pass host-side persistent buffers, and also explicitly avoids certain D2H transfers.

Thats effectively what this PR does; the CPU buffers in `CommonAttentionMetadata` are views into the `gpu_model_runner`s persistent `input_batch` host side tensors.