@LucasWilkinson 2025-07-23
@mgoin looks good 👍 I think we should land this since its a win and I can follow up if using numpy helps
```
VLLM_LOGGING_LEVEL=INFO cVLLM_USE_TRTLLM_DECODE_ATTENTION=1 VLLM_ATTENTION_BACKEND=FLASHINFER_VLLM_V1 lm_eval   --model vllm   --model_args '{"pretrained": "meta-llama/Meta-Llama
-3-8B-Instruct"}'   --tasks gsm8k --batch_size auto
...
WARNING 07-23 11:40:01 [flashinfer.py:140] Using TRTLLM decode attention (auto-detected). 
...                                                 
vllm ({'pretrained': 'meta-llama/Meta-Llama-3-8B-Instruct'}), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7559|±  |0.0118|
|     |       |strict-match    |     5|exact_match|↑  |0.7574|±  |0.0118|
```